rename divider hero component and file name
