+++ src/pages/LandingPage/Dashboard.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import {TransactionsPreview} from "../Transactions/Transactions";
 import LedgerInfo from "../LedgerInfo/LedgerInfo";
 import Typography from "@mui/material/Typography";
-import DividerHero from "../../components/Divider";
+import DividerHero from "../../components/DividerHero";
 import HeadingSub from "../../components/HeadingSub";
 import HeaderSearch from "../layout/Search";
 import Box from "@mui/material/Box";
